Stabilize Address Element launcher tests
Committed-By-Agent: codex
